requirements.txt
streamlit
torch
torchvision
librosa
numpy
matplotlib
pillow
tensorflow
sklearn
warnings
